move get_work_group to common
